4
https://github.com/opentibiabr/optimized_forgottenserver/commit/7a2905c93218500e8c45f2b0cc27d04d236b9d1a#

use asio::

+++ src/lua/global/globalevent.h
@@ -46,7 +46,7 @@ class GlobalEvents final : public Scripts {
 
 	private:
 		GlobalEventMap thinkMap, serverMap, timerMap;
-		int32_t thinkEventId = 0, timerEventId = 0;
+		uint64_t thinkEventId = 0, timerEventId = 0;
 };
 
 constexpr auto g_globalEvents = &GlobalEvents::getInstance;
